# website/static/json/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "couic",
        "description": "A lightweight eBPF-powered network filtering solution controllable through a REST API",
        "license": {
            "name": "MIT OR Apache-2.0",
            "identifier": "MIT OR Apache-2.0"
        },
        "version": "0.5.12"
    },
...
    "paths": {
        "/v1/sets/reload": {
            "post": {
                "tags": [
                    "sets"
                ],
                "summary": "Reload all sets",
                "description": "Triggers a reload of all CIDR sets from the configured sets directory.",
                "operationId": "reload_sets",
                "responses": {
                    "201": {
                        "description": "Sets reloaded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SetsReloadResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/v1/sets/{policy}": {
            "get": {
                "tags": [
                    "sets"
                ],
                "summary": "List all sets for a policy",
                "description": "Returns a summary of all CIDR sets for the specified policy.",
                "operationId": "list_sets",
                "parameters": [
                    {
                        "name": "policy",
                        "in": "path",
                        "description": "Policy name (drop or ignore)",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Policy"
                        },
                        "example": "drop"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of set summaries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SetSummary"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid path parameter)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "sets"
                ],
                "summary": "Create a new set",
                "description": "Creates a new CIDR set for the specified policy.",
                "operationId": "create_set",
                "parameters": [
                    {
                        "name": "policy",
                        "in": "path",
                        "description": "Policy name (drop or ignore)",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Policy"
                        },
                        "example": "drop"
                    }
                ],
                "requestBody": {
                    "description": "Set to create",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Set created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Set"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (malformed JSON or invalid path parameter)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/v1/sets/{policy}/{name}": {
            "get": {
                "tags": [
                    "sets"
                ],
                "summary": "Get a specific set",
                "description": "Returns the full set with all entries for the specified policy and set name.",
                "operationId": "get_set",
                "parameters": [
                    {
                        "name": "policy",
                        "in": "path",
                        "description": "Policy name (drop or ignore)",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Policy"
                        },
                        "example": "drop"
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Set name",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/SetName"
                        },
                        "example": "malware-ips"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Set found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Set"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid path parameters)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Set not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "sets"
                ],
                "summary": "Update a set",
                "description": "Replaces all entries in the specified set.",
                "operationId": "update_set",
                "parameters": [
                    {
                        "name": "policy",
                        "in": "path",
                        "description": "Policy name (drop or ignore)",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Policy"
                        },
                        "example": "drop"
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Set name",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/SetName"
                        },
                        "example": "malware-ips"
                    }
                ],
                "requestBody": {
                    "description": "Updated set data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Set updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Set"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (malformed JSON or invalid path parameters)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Set not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "sets"
                ],
                "summary": "Delete a set",
                "description": "Deletes the specified set. A reload may be required to apply changes.",
                "operationId": "delete_set",
                "parameters": [
                    {
                        "name": "policy",
                        "in": "path",
                        "description": "Policy name (drop or ignore)",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Policy"
                        },
                        "example": "drop"
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Set name",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/SetName"
                        },
                        "example": "malware-ips"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Set deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SetDeleteResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid path parameters)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Set not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CompositeError": {
                "type": "object",
                "required": [
                    "code",
                    "message",
                    "errors"
                ],
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        },
                        "propertyNames": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "ErrorCode": {
                "type": "string",
                "description": "Enum representing error codes used in the API",
                "enum": [
                    "Eprocessing",
                    "Eunauthorized",
                    "Enotfound",
                    "Econflict",
                    "Ebadrequest",
                    "Einvalid",
                    "Einternal",
                    "Enotimplemented"
                ]
            },
            "ErrorDetail": {
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "Policy": {
                "type": "string",
                "description": "Firewall policy type",
                "enum": [
                    "drop",
                    "ignore"
                ]
            },
            "SetsReloadResponse": {
                "type": "object",
                "required": [
                    "reload_status"
                ],
                "properties": {
                    "reload_status": {
                        "type": "string"
                    }
                }
            },
...
            "SetName": {
                "type": "string",
                "description": "Set name (1-48 chars, alphanumeric with hyphens and underscores)",
                "minLength": 1,
                "maxLength": 48,
                "pattern": "^[a-zA-Z0-9_-]+$",
                "examples": [
                    "malware-ips",
                    "blocklist_v1"
                ]
            },
            "Set": {
                "type": "object",
                "description": "A CIDR set with name and entries",
                "required": [
                    "name",
                    "entries"
                ],
                "properties": {
                    "name": {
                        "$ref": "#/components/schemas/SetName"
                    },
                    "entries": {
                        "type": "array",
                        "description": "List of CIDR entries in the set",
                        "items": {
                            "type": "string",
                            "description": "CIDR notation (IPv4 or IPv6)",
                            "examples": [
                                "192.168.1.0/24",
                                "2001:db8::/32"
                            ]
                        }
                    }
                }
            },
            "SetSummary": {
                "type": "object",
                "description": "Summary information for a set (used in list operations)",
                "required": [
                    "name",
                    "entry_count",
                    "file_size"
                ],
                "properties": {
                    "name": {
                        "$ref": "#/components/schemas/SetName"
                    },
                    "entry_count": {
                        "type": "integer",
                        "minimum": 0,
                        "description": "Number of entries in the set"
                    },
                    "file_size": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0,
                        "description": "Size of the set file in bytes"
                    }
                }
            },
            "SetRequest": {
                "type": "object",
                "description": "Request body for creating or updating a set",
                "required": [
                    "name",
                    "entries"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Set name (1-48 chars, alphanumeric with hyphens and underscores)",
                        "minLength": 1,
                        "maxLength": 48,
                        "pattern": "^[a-zA-Z0-9_-]+$",
                        "examples": [
                            "malware-ips"
                        ]
                    },
                    "entries": {
                        "type": "array",
                        "description": "List of CIDR entries for the set",
                        "items": {
                            "type": "string",
                            "description": "CIDR notation (IPv4 or IPv6)",
                            "examples": [
                                "192.168.1.0/24",
                                "10.0.0.0/8"
                            ]
                        }
                    }
                }
            },
            "SetDeleteResponse": {
                "type": "object",
                "description": "Response when deleting a set",
                "required": [
                    "deleted",
                    "reload_required"
                ],
                "properties": {
                    "deleted": {
                        "type": "boolean",
                        "description": "Whether the set was successfully deleted"
                    },
                    "reload_required": {
                        "type": "boolean",
                        "description": "Whether a reload is required to apply changes"
                    }
                }
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Uuidv4",
                "description": "An API key is a token that you provide when making API calls. Include the token in a header parameter called `Authorization`.\nExample: `Authorization: Bearer b03dd133-3827-4793-b0df-37f6ffbdd9a9`"
            }
        }
    },
...
}